+++ src/hooks/useAuth.jsx
@@ -1,6 +1,5 @@
-import { createContext, useContext, useState, useEffect } from 'react';
-import { currentUser } from '../data/userData';
-const AuthContext = createContext();
+import { useContext } from 'react';
+import AuthContext from '../contexts/AuthContext';
 
 export const useAuth = () => {
   const context = useContext(AuthContext);
@@ -8,48 +7,4 @@ export const useAuth = () => {
     throw new Error('useAuth must be used within an AuthProvider');
   }
   return context;
-};
-
-export const AuthProvider = ({ children }) => {
-  const [user, setUser] = useState(null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    // 페이지 새로고침 시 로컬 스토리지에서 사용자 정보 확인
-    const savedUser = localStorage.getItem('user');
-    if (savedUser) {
-      setUser(JSON.parse(savedUser));
-    }
-    setLoading(false);
-  }, []);
-
-  const login = async (username, password) => {
-    // API 호출을 통해 로그인 처리
-    if (username === 'admin' && password === 'password') {
-      setUser(currentUser);
-      localStorage.setItem('user', JSON.stringify(currentUser));
-      return currentUser;
-    } else {
-      throw new Error('Invalid credentials');
-    }
-  };
-
-  const logout = () => {
-    setUser(null);
-    localStorage.removeItem('user');
-  };
-
-  const value = {
-    user,
-    login,
-    logout,
-    loading,
-    isAuthenticated: !!user
-  };
-
-  return (
-    <AuthContext.Provider value={value}>
-      {children}
-    </AuthContext.Provider>
-  );
 };
